add topdocs info directly in the staticprops function
